Move `app.py` contents to `old_app.py`
